Rename namespace since it does not include any dispatching logic
