+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Integration/Helpers/IntegrationEventTypeFactory.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System;
-using Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.EnergySupplierChange;
+using Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEvents.EnergySupplierChange;
 
 namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.Helpers
 {
