+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Integration/IntegrationEvents/EnergySupplierChange/EnergySupplierChangedTopic.cs
@@ -12,14 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching
+namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEvents.EnergySupplierChange
 {
-    /// <summary>
-    /// An event that published outside the domain and bounded context
-    /// </summary>
-    #pragma warning disable CA1040
-    public interface IIntegrationEvent
-    {
-    }
-    #pragma warning restore
+    public record EnergySupplierChangedTopic(string Name) : Topic;
 }
